+++ agent/pubsub/pubsub.go
@@ -27,7 +27,7 @@ const (
 
 var (
 	pubsubPrefix             = flag.String("pubsub-prefix", "", "Prefix of Pub/Sub topics and subscriptions names.")
-	maxPubSubLeaseExtenstion = flag.Duration("pubsub-lease-extension", 0, "The max duration to extend the leases for a Pub/Sub message. If 0, will use the default Pub/Sub client value (10 mins).")
+	maxPubSubLeaseExtenstion = flag.Duration("pubsub-lease-extension", 20*time.Minute, "The max duration to extend the leases for a Pub/Sub message. If 0, will use the default Pub/Sub client value (10 mins).")
 	copySubGoroutines        = flag.Int("threads", 100, "The number of goroutines that serve the copy subscription. If 0, will use the default Pub/Sub client value (1000).")
 )