+++ CHANGELOG.md
@@ -10,6 +10,9 @@ A patch version bump is used for any change that does not affect the supported r
 major job run versions.
 
 ## [Unreleased]
+### Changed
+- Make the incremental copy chunk size flag controlled, default to 128MiB.
+  Increase the default PubSub message timeout to 20m.
 
 ## [1.0.8] - 2019-05-24
 ### Added